Create Predicting the term deposit subscription for Portuguese Banking Institution
The data taken for our project is one such dataset that belongs to direct marketing through phone calls
where the main objective is to classify the potential customers who can subscribe to the term deposit.
This belongs to a Portuguese bank dated from May 2008 to November 2010. The main purpose of this
project is to create a predictive machine learning model, using various supervised classification models
and identifying the best model that can classify the potential customers from the dataset who might be
able to subscribe to the term deposit based on the given features. This will in turn help to find the
features which have higher impact on the outcome